# src/main/java/com/edutrack/domain/user/dto/CompleteSignupRequest.java
package com.edutrack.domain.user.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class CompleteSignupRequest {

  private String signupToken;

}