# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: EduTrack API
  version: 1.0.0
  description: EduTrack 백엔드(Spring Boot 3, Java 17)용 OpenAPI 스펙
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/auth/signup/request:
    post:
      tags: [Auth]
      summary: 회원가입
      operationId: signupRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '200':
          description: 가입 신청이 완료되었습니다. 이메일 인증을 해주세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupInitResponse'
  /api/auth/send-email-verification:
    post:
      tags: [Auth]
      summary: 이메일 인증 코드 발송
      operationId: sendEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailVerificationRequest'
      responses:
        '200':
          description: 인증 코드가 발송되었습니다.
          content:
            application/json:
              schema:
                type: string
  /api/auth/verify-email:
    post:
      tags: [Auth]
      summary: 이메일 인증 확인
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: 이메일 인증이 완료되었습니다.
          content:
            application/json:
              schema:
                type: string
  /api/auth/academy-verify:
    post:
      tags: [ Auth ]
      summary: 학원 코드 인증
      operationId: verifyAcademy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademyVerifyRequest'
      responses:
        '200':
          description: 학원 코드 인증이 완료되었습니다.
          content:
            application/json:
              schema:
                type: string
  /api/auth/signup/complete:
    post:
      tags: [Auth]
      summary: 최종 회원가입 완료
      operationId: completeSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteSignupRequest'
      responses:
        '200':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'

  ########################
  # QuestionStatisticsController
  ########################
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignupRequest:
      type: object
      properties:
        loginId:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        academyCode:
          type: string
      required: [loginId, password, name, phone, email, academyCode]
    SignupInitResponse:
      type: object
      properties:
        signupToken:
          type: string
      required: [signupToken]
    SignupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loginId:
          type: string
        name:
          type: string
        role:
          type: string
    SendEmailVerificationRequest:
      type: object
      properties:
        signupToken:
          type: string
      required: [signupToken]
    VerifyEmailRequest:
      type: object
      properties:
        signupToken:
          type: string
        inputCode:
          type: string
      required: [signupToken, inputCode]
    AcademyVerifyRequest:
      type: object
      properties:
        signupToken:
          type: string
        academyCode:
          type: string
      required: [signupToken, academyCode]
    CompleteSignupRequest:
      type: object
      properties:
        signupToken:
          type: string
      required: [signupToken]
...